+++ src/org/cicirello/search/concurrent/ParallelMultistarterUtil.java
@@ -21,9 +21,11 @@
 package org.cicirello.search.concurrent;
 
 import org.cicirello.search.ReoptimizableMetaheuristic;
+import org.cicirello.search.Metaheuristic;
 import org.cicirello.search.ProgressTracker;
 import org.cicirello.search.problems.Problem;
 import org.cicirello.search.restarts.ReoptimizableMultistarter;
+import org.cicirello.search.restarts.Multistarter;
 import org.cicirello.search.restarts.RestartSchedule;
 import org.cicirello.util.Copyable;
 import java.util.ArrayList;
@@ -37,7 +39,7 @@ import java.util.Iterator;
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
- * @version 1.25.2021
+ * @version 3.21.2021
  */
 final class ParallelMultistarterUtil {
 	
@@ -45,6 +47,64 @@ final class ParallelMultistarterUtil {
 	 * Strictly a utility class, so default constructor is private.
 	 */
 	private ParallelMultistarterUtil() {}
+	
+	/**
+	 * Creates a list of Multistarters.
+	 *
+	 * @param search A Metaheuristic
+	 * @param schedules A collection of RestartSchedules
+	 * @return a list of Multistarters, one for each restart schedule,
+	 * all with identical and independent copies of search
+	 * @throws IllegalArgumentException if the collection of schedules is empty
+	 */
+	static <T extends Copyable<T>> Collection<Multistarter<T>> toMultistarters(Metaheuristic<T> search, Collection<? extends RestartSchedule> schedules) {
+		if (schedules.size() < 1) throw new IllegalArgumentException("Must pass at least one schedule.");
+		ArrayList<Multistarter<T>> restarters = new ArrayList<Multistarter<T>>(schedules.size());
+		boolean addedFirst = false;
+		for (RestartSchedule r : schedules) {
+			if (addedFirst) restarters.add(new Multistarter<T>(search.split(), r));
+			else {
+				restarters.add(new Multistarter<T>(search, r));
+				addedFirst = true;
+			}
+		}
+		return restarters;
+	}
+	
+	/**
+	 * Creates a list of Multistarters.
+	 *
+	 * @param searches A collection of Metaheuristics
+	 * @param schedules A collection of RestartSchedules
+	 * @return a list of Multistarters, such that multistarter i gets 
+	 * metaheuristic i and restart schedule i.
+	 * @throws IllegalArgumentException if searches.size() is not equal to schedules.size()
+	 * @throws IllegalArgumentException if not all metaheuristics solve the same problem
+	 * @throws IllegalArgumentException if the metaheuristics don't all share a single ProgressTracker
+	 */
+	static <T extends Copyable<T>> Collection<Multistarter<T>> toMultistarters(Collection<? extends Metaheuristic<T>> searches, Collection<? extends RestartSchedule> schedules) {
+		if (searches.size() != schedules.size()) {
+			throw new IllegalArgumentException("number of searches and number of schedules must be the same");
+		}
+		ArrayList<Multistarter<T>> restarters = new ArrayList<Multistarter<T>>(searches.size());
+		Iterator<? extends RestartSchedule> rs = schedules.iterator();
+		ProgressTracker<T> t = null; 
+		Problem<T> problem = null;
+		for (Metaheuristic<T> s : searches) {
+			if (problem == null) {
+				problem = s.getProblem();
+			} else if(s.getProblem() != problem) {
+				throw new IllegalArgumentException("All Metaheuristics in searches must solve the same problem.");
+			}
+			if (t==null) {
+				t = s.getProgressTracker();
+			} else if (s.getProgressTracker() != t) {
+				throw new IllegalArgumentException("All Metaheuristics in searches must share a single ProgressTracker.");
+			}
+			restarters.add(new Multistarter<T>(s, rs.next()));
+		}
+		return restarters;
+	}
 	
 	/**
 	 * Creates a list of ReoptimizableMultistarters.